refactor: [FE] createPortal을 통해 모달을 띄우도록 처리 createPortal을 통해 모달을 띄우도록 처리 (#118)
