+++ frontEnd/src/components/modal/Modals.tsx
@@ -1,14 +1,18 @@
+import { createPortal } from 'react-dom';
 import useModalState, { MODAL_COMPONENTS } from '@/stores/useModalState';
 import Modal from './Modal';
+
+const portalElement = document.getElementById('modal') as HTMLElement;
 
 export default function Modals() {
   const { modals } = useModalState((state) => state);
   const { hideModal } = useModalState((state) => state);
-  return (
+  return createPortal(
     <>
       {modals.map((id) => (
         <Modal hide={() => hideModal(id)} key={id} Component={MODAL_COMPONENTS[id].Component} modalProps={MODAL_COMPONENTS[id].props} />
       ))}
-    </>
+    </>,
+    portalElement,
   );
 }
