+++ frontEnd/src/main.tsx
@@ -3,7 +3,7 @@ import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import '@styles/index.css';
 import Home from '@pages/Home.tsx';
 import Room from '@pages/Room.tsx';
-import Modals from './components/common/Modals';
+import Modals from './components/modal/Modals';
 
 const router = createBrowserRouter([
   {
@@ -17,10 +17,10 @@ const router = createBrowserRouter([
 ]);
 function Main() {
   return (
-    <div>
+    <>
       <RouterProvider router={router} />
       <Modals />
-    </div>
+    </>
   );
 }
 ReactDOM.createRoot(document.getElementById('root')!).render(<Main />);
